refactor: Major overhaul of gossip code and DR
Reworked the way our gossip protocol works and the parties for Double Ratchet. We can now do E2E encryption between two nodes in the network with no direct connection, through a network of peers between them.
Also added the basis of message signing that will be needed later on.